server_host/requirements.txt
# Server requirements for render.com deployment
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
aiohttp>=3.10.0
aiofiles>=24.1.0
python-multipart>=0.0.9
websockets>=12.0
email-validator>=2.0.0
jinja2>=3.1.2
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0  # Pin to 4.x to avoid passlib compatibility issues
pydantic>=2.0.0
pydantic-settings>=2.0.0

aiofiles>=23.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
xxhash>=3.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
faker>=19.0.0